perf(web): lazy load detail navigation panels
